refactor(page-alert): create single source of truth for PageAlertType
Moves PageAlertType values into a dedicated types file using the
same pattern as ontario-button.types.ts.

- Adds PageAlertTypes constant array
- Derives PageAlertType from the array
- Removes duplicate validTypes definition from the component

This ensures runtime validation and TypeScript types stay in sync.
