+++ packages/ontario-design-system-component-library/src/components/ontario-page-alert/ontario-page-alert.types.ts
@@ -1,0 +1,3 @@
+export const PageAlertTypes = ['informational', 'warning', 'success', 'error'] as const;
+
+export type PageAlertType = (typeof PageAlertTypes)[number];
